+++ game/fx/creature.go
@@ -248,6 +248,8 @@ func Creature(card *match.Card, ctx *match.Context) {
 						ctx.Match.EndWait(card.Player)
 						ctx.Match.CloseAction(opponent)
 
+						ctx.Match.HandleFx(match.NewContext(ctx.Match, &match.AttackConfirmed{CardID: card.ID, Player: true, Creature: false}))
+
 						if len(shieldzone) < 1 {
 							// Win
 							ctx.Match.End(card.Player, fmt.Sprintf("%s won the game", ctx.Match.PlayerRef(card.Player).Socket.User.Username))