+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-...
+- Removed doublebreaker from "Supporting Tulip"
 
 ## [v2.2] - 21/01/2022