+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Corrected Muramasa, Duke of Blades' power from 3000 to 2000
 - Fixed an issue where Silver Axe kept adding mana when clicking "attack creature" while opponent had no creatures in the battle zone
 - Added new filter "race" in deck builder
+- Add button to stop spectating
 
 ## [v2.2] - 21/01/2022